elearning/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.3
cffi==1.17.1
channels==3.0.4
channels_redis==4.2.1
constantly==23.10.4
crispy-tailwind==1.0.3
cryptography==44.0.2
daphne==3.0.2
Django==4.2.19
django-admin-interface==0.29.4
django-bootstrap4==24.4
django-colorfield==0.12.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-tailwind==3.6.0
djangorestframework==3.15.2
drf-yasg==1.21.9
factory_boy==3.3.3
Faker==35.2.2
graphviz==0.20.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
msgpack==1.0.8
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydot==3.0.4
pyOpenSSL==25.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
service-identity==24.2.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
text-unidecode==1.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
zope.interface==7.2
